fix: all the fixes but not all
