+++ package.json
@@ -63,7 +63,7 @@
 		"@sapphire/async-queue": "^1.0.1",
 		"@sapphire/decorators": "^2.0.3",
 		"@sapphire/discord.js-utilities": "^1.3.1",
-		"@sapphire/plugin-api": "^1.3.1",
+		"@sapphire/plugin-api": "^1.5.1",
 		"@sapphire/plugin-i18next": "^1.2.1",
 		"@sapphire/snowflake": "^1.2.4",
 		"@sapphire/stopwatch": "^1.0.4",
