+++ src/lib/util/constants.ts
@@ -1,9 +1,11 @@
 /* eslint-disable @typescript-eslint/no-namespace */
+import { transformOauthGuildsAndUser } from '#lib/api/utils';
 import { CATEGORIES as TRIVIA_CATEGORIES } from '#lib/games/TriviaManager';
 import { Logger } from '#lib/logger/Logger';
 import { Colors, LanguageFormatters } from '#lib/types/Constants';
 import { DEV, VERSION as SKYRA_VERSION } from '#root/config';
 import { getHandler } from '#root/languages/index';
+import { ServerOptionsAuth } from '@sapphire/plugin-api';
 import { codeBlock, toTitleCase } from '@sapphire/utilities';
 import type { ClientOptions } from 'discord.js';
 import i18next, { FormatFunction } from 'i18next';
@@ -297,6 +299,11 @@ export const clientOptions: Partial<ClientOptions> = {
 	},
 	logger: {
 		instance: new Logger({ level: DEV ? LogLevel.Debug : LogLevel.Info })
+	},
+	api: {
+		auth: ({
+			transformers: [transformOauthGuildsAndUser]
+		} as unknown) as ServerOptionsAuth
 	},
 	i18n: {
 		defaultMissingKey: 'default',
