+++ src/routes/banners.ts
@@ -1,5 +1,5 @@
+import { ratelimit } from '#lib/api/utils';
 import { DbSet } from '#lib/database';
-import { ratelimit } from '#utils/util';
 import { ApplyOptions } from '@sapphire/decorators';
 import { ApiRequest, ApiResponse, methods, Route, RouteOptions } from '@sapphire/plugin-api';
