+++ src/routes/users/user/settings.ts
@@ -1,5 +1,5 @@
+import { authenticated, ratelimit } from '#lib/api/utils';
 import { DbSet } from '#lib/database';
-import { authenticated, ratelimit } from '#utils/util';
 import { ApplyOptions } from '@sapphire/decorators';
 import { ApiRequest, ApiResponse, methods, Route, RouteOptions } from '@sapphire/plugin-api';
 import { inspect } from 'util';
